BUG: Fixing z-score calculation segfault
There was a segmentation fault resulting from z-score calculation when a simple
mean (instead of a robust mean) is calculated. This was the result of array
that was not initialized properly. There were also some simple white-space
fixes and the addition of a counter to keep the progress messages from being
overly verbose. This reduces the progress output by a factor of 100.