BUG: Missing VTK_DIR in initial cache when re-building using a script
When ctest -S is run, we overwrite the current cmakecache.txt file so
that the values in the script are used.  However, any user-specified
or cmake-derived values from the initial run (e.g., VTK_DIR) are lost!
The current solution is hack-ish.  Ideally we should only modify the
necessary variables in the current cache, not replace all of them.