ENH: Add itkGradientBasedAngleOfIncidenceImageFilter.
Compute the cosine of the angle of beam incidence with the direction to the
ultrasound probe origin for a phased array or curvilinear probe using the
gradient of the input image.  This is computed with the scalar product of the
normalized gradient vector and vector from the location in space of the pixel to
the origin.

The type and instance of the filter used to compute the gradient can be
specified.

A tolerance is available on the gradient magnitude.

Tests were made with the acoustic impedance image as input, using the default
GradientImageFilter and a GradientRecursiveGaussianImageFilter, and with
different UltrasoundProbeOrigin's.