ENH: Support for linear arrays to itkGradientBasedAngleOfIncidenceImageFilter.
Linear array type ultrasound probes always have the same beam direction,
even if they are beam-steered, i.e. the beams are always parallel.

The itkGradientBasedAngleOfIncidenceImageFilter new has support for setting the
UltrasoundProbeType.  If the UltrasoundProbeType is CURVILINEAR or PHASED, then
the UltrasoundProbeOrigin is used to determine the beam direction on a
pixel-by-pixel basis.  If the UltrasoundProbeType is LINEAR, then the BeamDirection
is used for all pixels.