ENH: Parameter serializer.
A set of classes are added to serialize and deserialize parameters of
an itk::LightObject.  The base class, ParameterSerializer, is defined.
Serializer classes corresponding to other itk::LightObject's derive from
this class and define the parameters of the class and their default values.

Values are stores in a set of lightweight classes with well defined types.
These values are stored and retrieved with an Archiver class.  A JsonCppArchiver
was implemented to store and retrieve object parameters from a JSON file.

The common classes are stored in a separate TubeTK repository and built with
ExternalProject so that shared code can also be used with SlicerExecutionModel
without duplication.

A test case is created with the AcousticImpulseResponseImageFilterSerializer,
which is a case that included a parameter that is a virtual Object (the
gradient magnitude filter).  This tests the recursive serialization of Object
member parameters.

Change-Id: Ib8c36fc0c7298bba43603115e2f1d832e045d629