ENH: SlicerExecutionModel Module Serializer.
SlicerExecutionModel CLI's parameter can be serialized and deserialized to/from
a JSON file.  GenerateCLP is extended with a GenerateCLPSerializer program that
will store all the parameters and their descriptions to a JSON file specified
with the --parametersarchive flag or it can load all the parameter values with
the file specified with the --parameterstorestore flag.

Flags specified on the command line will override values restored from the JSON archive.

A custom version of SlicerExecutionModel that provides this functionality is
used in the standalone Superbuild.

Change-Id: I97695324bd1cd2bd297a7a3dc442fe84312ac7f1