BUG: Find the JsonCpp library location with better CMake target exports.
The JsonCpp, ParameterSerializer, and SlicerExecutionModel *Config.cmake
files were updated so CMake has better knowledge of where the json library
lives when building against them.  This is achieved by creating a *Targets.cmake
file with export(TARGETS... and install(EXPORTS... then making sure it is include'd
in the *Config.cmake file.  If a library in a project links against a library
in a project it depends on, the *Config.cmake for the project it depends on is include'd
in the dependent project's *Config.cmake

This helps CMake pass around the library locations in a way that is robust
to platform and CMAKE_BUILD_TYPE.

Change-Id: I31c182a326450095a0c1e3edeb22c7be6ee41093