ENH: Add class to record optimization parameter progression.
When set to observe a SingleValuedNonLinearOptimizer, it will record
the parameters, the cost function value, and the cost function derivative
at every iteration, then save the result to an HDF5 file.  An HDF5 file
is used so it can be easily read and examined in Python while retaining
and binary precision and compression.

Change-Id: I7737cd6951c32b7f2efe0d6393469fbc8c458037