ENH: Do not hard code Optimizer in ImageToTubeRigidRegistration.
Move from a hard-coded derived OptimizerType to a default derived Optimizer
(currently GradientDescent) for several advantages.  Since the default optimizer
is instantiatied in the constructor, ->GetOptimizer() will return the default
optimizer right after the object is contructored (needed for the next patch).
The setting of optimizer parameters is simplified.  If a different derived
Optimizer type is desired, the ImageToTubeRigidRegistration API does not have to
re-written and all the code that uses it changed.

Change-Id: I28abd8f93e4834e6d850b1f716f6c2b5f9f86f8e