ENH: Add new tubeMacro.h and very lightweight Object base class
My original thought was that everything could be in itk::tube and use
itkMacro.h. That idea was abandoned, so here is a cut down and
simplified version of itkMacro.h / vtkSetGet.h with helper macros for classes
that are not derived from ITK or VTK.

The Object base class simplifies printing out non-ITK or -VTK derived classes
for debugging. No smart pointers or anything heavyweight.

Change-Id: Ib710d42414b226bae3141fc53e084dcb6ac15511