ENH: Convert tubes to NumPy arrays.
Write a C Python extension module to convert the VesselTubeSpatialObjectPoint's
in a file that contains a VesselTubeSpatialObject into a NumPy record array.
Each property in the tube point is a field in the NumPy record array.  The
length of the array equals the number of points in the tube. Example code:

  from tubetk.numpy import tubes_from_file
  tubes = tubes_from_file(filename)
  print(tubes['Radius'])

To make importable:

1) Turn TubeTK_USE_NumPy ON in the CMake configuration.
2) Make a symlink to /path/to/TubeTK-build/Python/Base/tubetk.pth at $(python -m site --user-site).

Todo's:

1) Support 2D.
2) Do not include VesselTubeSpatialObject fields for a TubeSpatialObject.

Change-Id: If9e321e2a3f6d81644382917e8bbcac0536c6685