ENH: Fix PDF Segmenter to work with Prometheus
First, the PDF segmenter retreived the EditUtil module by calling the
slicer.modules.EditorWidget. It is fine as long as the editor isn't
integrated in another module. In that case, we don't want the integrated
editor to reference to the module in slicer.
Since Effect already have a accessor on EditUtil, we can simply refer to it
(the PDF segmenter is an effect).

Also, the PDF segmenter was using the command slicer.cli.run without
passing a CLI. This would create a new CLI everytime it was called.
Instead, we can create our own PDF segmenter.
A detail is that the normal PDF segmenter CLI node needs to be created
before the custom editor cli node. Otherwise the latter will be the one
 selected when opening the PDF segmenter CLI module.

Change-Id: I9b578d18c59a246ff6cc18ca73daa5a93c16791b