ENH: Visualize tube points as circles.
Visualize tube points as circles.  This captures the tube point size,
orientation, position, and an alpha value on the color can be used to indicate
the weight when performing registration, etc.

Tube points are visualized with pyqtgraph by creating pyqtgraph MeshData out
of circles.  The circles are created from a wagonwheel representation of
vertexes and faces.  A circle is created in the x-y plane, scaled with the
tube point Radius, rotated via the tube point Normal1, Normal2, and Tangent, and
translated using the tube point Position.

Change-Id: If04c6527f3acd666a5f8395576bbeb9b1c9f4d76