ENH: Enhance the Python testing virtualenv setup.
The virtualenv is created at build time, and successful results from previous
runs are cached for efficiency.

System site packages are used if possible.

Option variables are changed from TubeTK_TEST_EXAMPLES to TubeTK_USE_PYTHON,
and if that is enabled, TubeTK_USE_NOTEBOOKS, TubeTK_USE_NUMPY, and
TubeTK_USE_PYQTGRAPH are exposed.

Do not require setting the environmental variable in MergeAdjacentImage.ipyb.

The virtualenv is exposed at the top level so that the
PYTHON_TESTING_EXECUTABLE is also available to Base/ and Applications/.

Change-Id: I98ce8eb27a0ae96334f9fde74cac627ff862f722