COMP: Fix extension build simplifying Qt version requirement rules
See Slicer r22466 commit:

// --------------
As explained in issue #3325, on linux to prevent a crash occurring when
a page is loaded in a QWebView, at least Qt 4.8.1 is needed
on Ubuntu 13.04. Alberto also reported it is crashing when using
Ubuntu 12.04 (see #3394).

To prevent developer from having to explicitly pass
the Slicer_REQUIRED_QT_VERSION option, the requirement scheme
has been simplified.

Removed non working option "Slicer_EXPERIMENTAL_QT_VERSION" and requires
at least Qt 4.7.4 on all platform.

This revert commit r21743 and partially revert r21727

We will discuss with the community if it makes sens to required Qt >= 4.8.1
on all platform. On MacOSX, this implies that Leopard won't be supported
anymore.
// --------------