BUG: More sides to tube SpatialObjectViewer rendering.
This applies and sets the number of sides to a higher value (32).  The default
could be decreased in the future.

This does not address of bug of setting the number of sides display property (which is
not getting applied).  The vtkTubeFilter default of 3 sides is only being
displayed without this patch.