ENH: Add ConvertTubesToSurface Application.
Convert a tubes .tre file (has points, radius, other descriptors at points)
into a file with a polydata representation of the tube surface.  Files can be
save in the .vtk legacy VTK file format or the .stl file format.  The .stl
file format can applied to 3D printers, for example.  The .vtk format also
retains some of the scalar data associated with the tubes.  It can be used to,
for example, visualize the tube Id's in Paraview and select sub-tube Id's.