BUG: Fix launcher PATH munging on Windows.
The launcher script is needed on Windows even with TubeTK_USE_SLICER OFF.

CMAKE_BUILD_TYPE is not used with Visual Studio builds. CMAKE_BUILD_TYPE is not
in the output path when using a Generator like Ninja. Make the path munging
include the common Debug and Release directories for Visual Studio and none for
Ninja (Ninja is to be tested).

Change-Id: Ib2c23e2c20df660844965d7428eca18222f7e20d