Fix TRE file loading in vesselview
The previous commit was allowing normal TubeSpatialObject files to be
loaded but broke compatibility with VesselTubeSpatialObject files.

The problem is that an itk::TubeSpatialObject<3> isn't the same
as a itk::VesselTubeSpatialObject<3>::Superclass (i.e. the parent class)
since VesselTubeSpatialObject changes the point type used.
While the node was accepting both kind of tube types they do not share the
same base class (at least from a casting point of view), making it hard to
do the same thing for both classes with a clean code.

To remediate this in a simple way, we decide to only support
VesselTubeSpatialObjects in a vtkMRMLSpatialObjectNode.

Consequently modified itktubeTubeXIO to make sure it generates
VesselTubeSpatialObjects compatible files.