Add option to save the tortuosity arrays as an CSV file
This implies a little bit of refactoring/enhancing:
 - Add All flag (0xFF) so we can run all the metrics
 - Refactor the GetArray(s) methods to clarify when the array is created
or not. To this end, a GetOrCreateArray method was added. Internaly, we
use templated method so we can any type of array.
  - Automatically add the NumberOfPoints integer array on each computation
of tortuosity. For each vessels, this array values tells the user how many
point this vessel has. It can be used (like in SaveAsCSV) to only get one
measure value per vessel instead of on measure value of each point of the
vessel.

By default, the csv does not list the InflectionPoints array since it only
makes sense to look at it when we look at all the points of the vessels.