Refactor the creation of display nodes for spatial object
It was done both in the logic the node two different ways. Instead, we
move all the display node creation in the node. We also enable tubes by
default.

Note that at this point, there are visualization issues still. The tube
aren't visible until we click on the node in the spatial object widget.
This will be addressed in the next commit.