COMP: Ensure TubeTK CLI build against Slicer register the ITKIO factories
This commit fixes the problem  avoiding forcing the CLI library wrapper
variable into the cache.

Forcing the values into the cache is not needed because

 (1) when TubeTK is built standalone it will use the default wrapper
provided by the SlicerExecutionModel

and

 (2) when TubeTk is build against Slicer, the SlicerConfig.cmake will
set the Slicer SEM wrapper variables in the current scope that
will then be picked up automatically by the following call to
find_package(SlicerExecutionModel)

As detailed in (1), this commit also remove the uneeded tubeTk CLI wrapper.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>