Fix crash when selecting a new spatial object after deletion
What happened was that the node was correctly updated. This caused the
spatial objects widget (that handle line, tube and glyph display) to
update. These object hold a reference to the display node of the previously
deleted node. Trying to update said node would cause the crash (the display
node was deleted).

Two approach can be considered for fixing this:
 - We could always get the display node from the node.
 - We could make sure that the display node is correctly updated before
looking at it for the widget

The latter approach was chosen since it creates the least amount of code
change. (However if in the future it proves unsatisfactory, it might be
better to refactor the code to always get the display node from the node).

Change-Id: I560e0f1e19851e5a8d52baf2285165b91784da08