Add option for running CLIs in slicer or not
To be able to run a CLI in Slicer, one must have the variable ITK_NO_IO_FACTORY_REGISTER_MANAGER
set to 1 before calling the include of the ITK use file. Since TubeTK CLIs
support both running within and without Slicer, there needs to be an option
for setting or not the ITK_NO_IO_FACTORY_REGISTER_MANAGER variable.
The TubeTK_RUN_MODULES_IN_SLICER does so. It is also automatically set
to true when the user decides to build the slicer modules.