ENH: Used the ArrayFire Library to do gaussian derivative filtering on GPU
Made template only libraries TubeTKFiltering, TubeTKRegistration, TubeTKSegmentation
and as interface libraries. I had to do this because the new ArrayFire filters
need to link to ArrayFire's Libraries and use ArrayFire_INCLUDE_DIRS. And i want
to abstract the ArrayFire stuff from any other targets that use the stuff in
TubeTKFiltering.

The add_library( <TARGET-NAME> INTERFACE ) is only supported from cmake 3.0.0. So
we will have to bump up the minimum version of TubeTK to 3.0.0 or higher.