WIP: Add Docker Files
Add Docker files in order to set up continuous testing. This commit enables
the user to create a docker image to welcome the project.
Using the build.sh and run.sh scripts located in the
Docker-TubeTK_BUILD_USING_SLICER_ON repository, we can build, run and
test the project.
This commit is tagged as WIP because the continuous integration testing
has to be triggered using CircleCI. CircleCI support will be added in the
next commit.