COMP: Fix configuration of SegmentTubeUsingMinimalPath with ITK <= 4.8.1
The problem was encountered when doing a clean build of VesselView ( also
reproducible when building TubeTK with ITK <= 4.8.1 and BUILD_TESTING
disabled ).

This commit fixes a regression introduced in aef4fae (ENH: git clone of
minimalpath extension source during configure). Since GITCOMMAND is only
set when CTest module is included, this commit uses GIT_EXECUTABLE
variable that is set in the superbuild and passed down to the inner build.

The configuration output has also been cleaned up and looks like this:

[...]
-- Configuring SEM CLI module: SegmentTubes
-- Cloning MinimalPathExtraction into '/home/jcfr/Projects/VesselView-Release/S-bld/TubeTK-build/MinimalPathExtraction'
-- Checking out MinimalPathExtraction '4f526802b0056258ca1036fbe7e94d53b1458716'
-- Configuring SEM CLI module: SegmentTubeUsingMinimalPath
[...]