ENH: Addition of a possibility to skip the evolutionary optimizer
Previously, the evolutionary optimizer was only skipped for the affine registration if a rigid registration was performed first. Now, there is an option to skip this step if the user wants. The evolutionary optimizer is also skipped for any registration that follows an initial registration (i.e. skipping at BSpline step if an affine registration is performed), following the scheme that was already implemented for affine transforms.