ENH: RegisterImages only supports 2D images if set in CMake
RegisterImages supports 2D images if RegisterImages_Support_2D_Images if set to ON. This speeds up compilation time and reduces the size of the executable. By default, this variable is set to OFF. A new flag has been added to RegisterImages to allow a user to test if RegisterImages has been built with 2D image support. Tests have been added to verify that this new flag works.