ENH: Activation of 2D images support from CMake
All modules that use "tubeCLIHelperFunctions.h" can load and process 2D and 3D images. To avoid always compiling with 2D image support, an option has been added in CMake (TubeTK_Support_2D_Images). If this option is set to OFF (default), the modules only support 3D images. They may be able to process 2D images by loading them as 3D images with only one slice. It is not garanteed that the processing will be performed correctly and some ITK filters will throw exceptions if the input image does not have a required thickness (e.g. Gaussian Smoothing Filters). When TubeTK_Support_2D_Images is set to ON, the executables will be named ${MODULE_NAME}2D3D to let the user know that it handles both 2D and 3D images if the version of SlicerExecutionModel that the modules are compiled against handles the parameter "DISABLE_DEFAULT_CXX_FILE".