COMP: Build all TubeTK libraries into the same directory
This commit sets default output directories using TubeTK install directory.

Prior to this commit, libraries were built in both "<CMAKE_BINARY_DIR>/lib"
and "<CMAKE_BINARY_DIR>/lib/TubeTK".  After this commit, all libraries
are built into "<CMAKE_BINARY_DIR>/<TubeTK_INSTALL_LIB_DIR>"

Here is an example of output before this commit:

$ cd ./lib
$ ls -1
libAtlasBuilderUsingIntensityLib.so
libComputeBinaryImageSimilarityMetricsLib.so
libComputeImageSimilarityMetricsLib.so
[...]
libSegmentConnectedComponentsUsingParzenPDFsLib.so
libSegmentTubesLib.so
libSegmentTubeUsingMinimalPathLib.so
libSegmentUsingOtsuThresholdLib.so
libShrinkImageLib.so
libSimulateAcquisitionArtifactsUsingPriorLib.so
libtubeBaseCLITestsLib.a
libtubeBaseCommonTestsLib.a
[...]
libtubeBaseSegmentationTestsLib.a
libtubeBaseUSTKTestsLib.a
libTubeMathLib.so

$ cd ./lib/TubeTK
$ ls -1
libitkMinimalPathExtraction-4.10.so
libitkMinimalPathExtraction-4.10.so.1
libTubeTKCommon.a
libTubeTKIO.a
libTubeTKNumerics.a
libTubeTKObjectDocuments.a
libTubeTKUSTK.a