STYLE: Move implementation of CropFilter  into TukeTK/Base
The ITKModules/ITKTubeTK directory contains now classes that
have an instance of the Base class they are shadowing.

Renames itktubeCropImage to tubeCropImage

To avoid changing the access modifier associated with method like PrintSelf,
itkTubeTK class are declared as friend within the corresponding base class.

As discussed with Stephen, the adopted convention is the following:

* Classes in TubeTK/Base will either be in "itk:tube" or "tube" namespace.

* Classes in TubeTK/ITKModules/ITKTubeTK will always be in "tube" namespace.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>