STYLE: Facilitate re-use of ImageMath functionality
To avoid duplicating the code when integrating tubeTK pipeline expected to
following CLI convention, this commit introduces "ImageMath" static class
containing the functions listed below. In a follow up commit, more functions
will be moved from the application to the class.

Note that this commit do not change the original behavior. For example, error
management (boolean vs exception) and potential issues will be addressed in
follow-up commits.

List of functions:

* ApplyIntensityWindowing
* ApplyIntensityMultiplicativeWithBiasCorrection
* ResampleImage
* AddUniformNoise
* AddGaussianNoise
* AddImages
* MultiplyImages
* NormalizeImage
* FuseImages
* ThresholdImage
* ComputeImageStdDevOrMeanWithinRangeUsingMask