ENH: Update creation of python virtual environment
To be able to run the tests written in python, the configuration
process creates a python virtual environment in which it installs
the required packages.
Conditions have been added to the creation of the virtual
environment to avoid potential problems or misunderstandings
(starting from an already existing virtual environment, or
using the wrong virtualenv script). Error messages are printed
if missing python packages will make compilation fail. Warning
messages are printed if missing packages are likely to create
errors when running Python tests in the automatically created
virtual environment.