ENH: CACHE variables should not always be removed
When building TubeTK as a Slicer extension withing Slicer (e.g. as
part of VesselView), CACHE variables that are passed to TubeTK
should not be unset as they are not reset by SlicerConfig.cmake.
Indeed, in this case, SlicerConfig.cmake does not contain values
for these CACHE variables. The SlicerConfig.cmake file used in
this case is not the one contained in the build directory of Slicer,
but the one contained in the subdirectory containing the extensions.