ENH: Set environment variables for Python tests
When running python scripts that use TubeTK wrapped in Python,
one needs to set their environment so that the libraries are loaded.
This can be done two ways:
-Copy WrapITK.pth from TubeTK-build/Generators/Python/ or setting
ITK_BUILD_DIR and TubeTK_BUILD_DIR environment variables. Because tests
are develop to verify that TubeTK works correctly, they replicate the
expected behavior of a user. Python tests set these two environment
variables before running the tests. This is done by calling the
new TubeTK function Midas3FunctionWithEnvAddTest.
Additionally, a test has been added to verify that the environment
variables are set on the system and match the expected value
of these variables in this project. This test can be deactivated
by setting TubeTK_TEST_ENVIRONMENT to OFF (e.g. on a Dashboard
machine compiling multiple versions of TubeTK).
on the current computer, they should behave the way a user
would expect their own too