COMP: CircleCI compiler crash because container runs out of memory
CircleCI build are constrained by a 4Gb memory limit limit. See
https://circleci.com/docs/oom/

The compiler checks for the number of cores available on the machine.
The output of "lscpu" shows 32 cores on the entire machine. However,
builds are running in a single container that shares cores with other
currently running containers.
The output of "top" shows that the compiler try to use all cores
and can dedicate up to 1Gb of memory for each job :

  PID  USER      PR  NI  VIRT  RES  SHR S %CPU %MEM    TIME+  COMMAND
 17714 root      20   0  622m 578m  11m R   12  0.2   0:42.30 cc1plus
 17797 root      20   0  692m 647m  11m R   12  0.3   0:39.37 cc1plus
 [...]
 17570 root      20   0  840m 791m  11m R    4  0.3   0:46.91 cc1plus
 17723 root      20   0  729m 682m  10m R    4  0.3   0:38.86 cc1plus
 17864 root      20   0  664m 626m 6556 R    4  0.3   0:32.73 cc1plus

The 4Gb limit is reached quickly and jobs are randomly killed, throwing
the following output:
  c++: internal compiler error: Killed (program cc1plus)

Even with argument -j1, Ninja is still using all cores.
SITE_CMAKE_GENERATOR has been changed to "Unix Makefiles" and
SITE_MAKE_COMMAND has been changed to "make -j3" to limit memory usage.

${SITE_BUILD_TYPE} is now specified by circle.yml in order to fix the
build directory that was mounted on the machine. It was not reachable
because of name inconsistency.
Default is set to Release.