BUG: TubeTK_TEST_ENVIRONMENT not initialized
TubeTK_TEST_ENVIRONMENT was initialized only if TubeTK_TEST_ENVIRONMENT
was set to ON (if statement). However, since on the dashboards we
were trying to set it to OFF (ON by default), the value passed
by the dashboard configuration scripts was never taken into account.
The "if" statement was remove so that the value of
TubeTK_TEST_ENVIRONMENT is always initialized.