BUG: Compares absolute paths for ITK_BUILD_DIR and TubeTK_BUILD_DIR
Comparing the expected value of ITK_BUILD_DIR (respectively TubeTK_BUILD_DIR)
with the corresponding environment variable might not pass when comparing
the strings of the paths if the paths are not set in the same format
or if a path is relative and the other one absolute even if these
paths correspond to the same folder. On Windows, an additional problem
could come from the path in the environment variables being set
with back-slashes (Windows format) whereas CMake paths use forward-slashes
(UNIX format).