BUG: Add paths associated with MinimalPathExtraction module
When an ITK module is used, it might have been built as a remote module.

Include directories of remote modules are given a module-specific cmake file
that is included when ITK is included in a project, but the paths
for that module are not included when ITK include paths are included.
The paths of a remote module must be specifically included by applications
that use them.