ENH: Remove superbuild of ITK and VTK.
Only supported compilation of TubeTK now requires ITK (and optionally
VTK) to be already compiled.   Given the standard (and preferred)
practice of wrapping those toolkits (and TubeTK) for python as well
as the standard and preferred deployment of TubeTK as an ITK module,
the use of superbuild is infrequent and not recommended.