BUG: TubeTK STATIC, even if itkTubeTK SHARED (#167)
itkTubeTK is a two-library system:

1) The include directory contains the itkTubeTK library which depends
on the TubeTK library. The itkTubeTK library is a high-quality ITK interface
(that can be wrapped for python) to the TubeTK library.

2) The src directory contains the TubeTK library which is an (ancient)
mix of ITK and non-ITK classes, functions, applications, and
examples.   It is poorly documented and inconsistent in style, but it
is rich in functionality :)

Regretfully the old TubeTK does not support being built as a shared
library.  So, even if itkTubeTK is being built shared, the underlying
TubeTK library will always be built static.

* BUG: TubeTK STATIC, even if itkTubeTK SHARED

itkTubeTK is a two-library system:

1) The include directory contains the itkTubeTK library which depends
on the TubeTK library. The itkTubeTK library is a high-quality ITK interface
(that can be wrapped for python) to the TubeTK library.

2) The src directory contains the TubeTK library which is an (ancient)
mix of ITK and non-ITK classes, functions, applications, and
examples.   It is poorly documented and inconsistent in style, but it
is rich in functionality :)

Regretfully the old TubeTK does not support being built as a shared
library.  So, even if itkTubeTK is being built shared, the underlying
TubeTK library will always be built static.